# Conflicts:
#	modules/core/core-content/src/test/java/com/enonic/xp/core/impl/content/CreateContentCommandTest.java

+++ modules/core/core-api/src/main/java/com/enonic/xp/export/NodeExportListener.java
@@ -2,7 +2,7 @@ package com.enonic.xp.export;
 
 public interface NodeExportListener
 {
-    void nodeExported( long count );
+    void nodeExported( int count );
 
-    void nodeResolved( long count );
+    void nodeResolved( int count );
 }
